+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # storybook
 
+## 1.0.58
+
+### Patch Changes
+
+- update
+
 ## 1.0.57
 
 ### Patch Changes