+++ packages/component/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @orderly.network/components
 
+## 0.0.72
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.68
+  - @orderly.network/types@0.0.71
+  - @orderly.network/utils@0.0.59
+
 ## 0.0.71
 
 ### Patch Changes