+++ packages/futures/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @orderly.network/futures
 
+## 0.0.72
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.71
+  - @orderly.network/utils@0.0.59
+
 ## 0.0.71
 
 ### Patch Changes